pyproject.toml
[project]
name = "gitpulse-unraveling-open-source-evolution"
version = "0.1.0"
description = "Tracking open-source evolution through GitHub Archive data"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "db-dtypes>=1.4.4",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-bigquery-storage>=2.0.0",
    "google-cloud-storage>=2.0.0",
    "pandas>=2.0.0,<2.3.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "dbt-core>=1.8.0",
    "dbt-bigquery>=1.8.0",
    "streamlit>=1.55.0",
    "plotly>=6.6.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[dependency-groups]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.4.0",
]
